extract exec observable into lib
