+++ tools/__tasks__/test/javascript/index.js
@@ -1,18 +1,4 @@
-const execa = require('execa');
-const split = require('split');
-// eslint-disable-next-line import/no-unassigned-import
-require('any-observable/register/rxjs-all');
-const Observable = require('any-observable');
-const streamToObservable = require('stream-to-observable');
-
-const exec = (cmd, args) => {
-    const cp = execa(cmd, args);
-
-    return Observable.merge(
-        streamToObservable(cp.stdout.pipe(split()), { await: cp }),
-        streamToObservable(cp.stderr.pipe(split()), { await: cp })
-    ).filter(Boolean);
-};
+const exec = require('../../lib/exec-observable');
 
 const legacyTests = ['common', 'facia'].map(set => ({
     description: `Run ${set} tests (legacy)`,
